+++ BitmovinAnalyticsCollector/Classes/BitmovinPlayer/Classes/BitmovinPlayerAdapter.swift
@@ -7,15 +7,14 @@ class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
     private var errorCode: Int?
     private var errorMessage: String?
     private var drmPerformanceInfo: DrmPerformanceInfo?
-    private var isPlayingAd: Bool
     private var isStalling: Bool
     private var isSeeking: Bool
-    private var drmLoadTimeMs: Int64?
+    /// DRM certificate download time in milliseconds
+    private var drmCertificateDownloadTime: Int64?
 
     init(player: BitmovinPlayer, config: BitmovinAnalyticsConfig, stateMachine: StateMachine) {
         self.player = player
         self.config = config
-        self.isPlayingAd = false
         self.isStalling = false
         self.isSeeking = false
         super.init(stateMachine: stateMachine)
@@ -112,10 +111,10 @@ class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
 
         eventData.audioLanguage = player.audio?.language
         
-        if (videoStartFailed) {
-            eventData.videoStartFailed = videoStartFailed
-            eventData.videoStartFailedReason = videoStartFailedReason ?? VideoStartFailedReason.unknown
-            resetVideoStartFailed()
+        if (stateMachine.videoStartFailed) {
+            eventData.videoStartFailed = stateMachine.videoStartFailed
+            eventData.videoStartFailedReason = stateMachine.videoStartFailedReason ?? VideoStartFailedReason.unknown
+            stateMachine.resetVideoStartFailed()
         }
     }
 
@@ -137,55 +136,38 @@ class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
             return Util.timeIntervalToCMTime(_: player.currentTime)
         }
     }
-    
-    override func setVideoStartTimer() {
-        if(isPlayingAd){
-            return
-        }
-        
-        super.setVideoStartTimer()
-    }
-    
-    @objc override func willEnterForegroundNotification(notification: Notification){
-        if(!didVideoPlay && didAttemptPlay && !isPlayingAd){
-            setVideoStartTimer()
-        }
-    }
 }
 
 extension BitmovinPlayerAdapter: PlayerListener {
     func onPlay(_ event: PlayEvent) {
-        setVideoStartTimer()
-        didAttemptPlay = true
-        if (!isSeeking && !isStalling) {
-            stateMachine.transitionState(destinationState: .playing, time: Util.timeIntervalToCMTime(_: player.currentTime))
-        } else if (isStalling && stateMachine.state != .seeking && stateMachine.state != .buffering) {
+        stateMachine.play(time: nil)
+        
+        if (isStalling && stateMachine.state != .seeking && stateMachine.state != .buffering) {
              stateMachine.transitionState(destinationState: .buffering, time: Util.timeIntervalToCMTime(_: player.currentTime))
         }
     }
     
     func onPlaying(_ event: PlayingEvent) {
-        clearVideoStartTimer()
-        didVideoPlay = true
+        if (!isSeeking && !isStalling) {
+            stateMachine.playing(time: Util.timeIntervalToCMTime(_: player.currentTime))
+        }
     }
 
     func onAdBreakStarted(_ event: AdBreakStartedEvent) {
-        clearVideoStartTimer()
-        isPlayingAd = true
+        stateMachine.transitionState(destinationState: .ad, time: currentTime)
     }
     
     func onAdBreakFinished(_ event: AdBreakFinishedEvent) {
-        isPlayingAd = false
+        stateMachine.transitionState(destinationState: .adFinished, time: currentTime)
     }
     
     func onPaused(_ event: PausedEvent) {
         isSeeking = false
-        stateMachine.transitionState(destinationState: .paused, time: Util.timeIntervalToCMTime(_: player.currentTime))
+        stateMachine.pause(time: currentTime)
     }
 
     func onReady(_ event: ReadyEvent) {
         self.isPlayerReady = true
-        transitionToPausedOrBufferingOrPlaying()
     }
 
     func onStallStarted(_ event: StallStartedEvent) {
@@ -205,15 +187,16 @@ extension BitmovinPlayerAdapter: PlayerListener {
     }
 
     func onDownloadFinished(_ event: DownloadFinishedEvent) {
-        let downloadTimeInMs = BitmovinPlayerUtil.ms(t: event.downloadTime)
+        let downloadTimeInMs = event.downloadTime.milliseconds
 
         switch event.downloadType {
         case BMPHttpRequestTypeDrmCertificateFairplay:
             // This request is the first that happens when initializing the DRM system
-            self.drmLoadTimeMs = downloadTimeInMs
+            self.drmCertificateDownloadTime = downloadTimeInMs
         case BMPHttpRequestTypeDrmLicenseFairplay:
-            self.drmLoadTimeMs = (self.drmLoadTimeMs ?? 0) + (downloadTimeInMs ?? 0)
-            self.drmPerformanceInfo = DrmPerformanceInfo(drmType: DrmType.fairplay, drmLoadTime: self.drmLoadTimeMs)
+            let drmLoadTimeMs = (self.drmCertificateDownloadTime ?? 0) + (downloadTimeInMs ?? 0)
+            self.drmPerformanceInfo = DrmPerformanceInfo(drmType: DrmType.fairplay, drmLoadTime: drmLoadTimeMs)
+            self.drmCertificateDownloadTime = nil
         default:
             return
         }
@@ -247,13 +230,17 @@ extension BitmovinPlayerAdapter: PlayerListener {
     func onError(_ event: ErrorEvent) {
         errorCode = Int(event.code)
         errorMessage = event.message
-        if (!didVideoPlay) {
-            setVideoStartFailed(withReason: VideoStartFailedReason.playerError)
+        if (!stateMachine.didStartPlayingVideo && stateMachine.didAttemptPlayingVideo) {
+            stateMachine.setVideoStartFailed(withReason: VideoStartFailedReason.playerError)
         }
         stateMachine.transitionState(destinationState: .error, time: Util.timeIntervalToCMTime(_: player.currentTime))
     }
 
     func transitionToPausedOrBufferingOrPlaying() {
+        if(!stateMachine.didStartPlayingVideo) {
+            return
+        }
+        
         if isStalling {
             // Player reports isPlaying=true although onStallEnded has not been called yet -- still stalling
             stateMachine.transitionState(destinationState: .buffering, time: Util.timeIntervalToCMTime(_: player.currentTime))
@@ -265,8 +252,8 @@ extension BitmovinPlayerAdapter: PlayerListener {
     }
     
     func onSourceWillUnload(_ event: SourceWillUnloadEvent) {
-        if (!didVideoPlay && didAttemptPlay) {
-            self.onPlayAttemptFailed(withReason: VideoStartFailedReason.pageClosed)
+        if (!stateMachine.didStartPlayingVideo && stateMachine.didAttemptPlayingVideo) {
+            stateMachine.onPlayAttemptFailed(withReason: VideoStartFailedReason.pageClosed, time: delegate.currentTime)
         }
     }